APEXMALHAR-2034 Creating Avro Module in contrib
