+++ pom.xml
@@ -215,7 +215,6 @@
     <module>contrib</module>
     <module>kafka</module>
     <module>examples</module>
-    <module>avro</module>
   </modules>
 
   <dependencies>
